requirements.txt
colorzero==2.0
gpiozero==1.6.2
paho-mqtt==1.6.1
psutil==5.8.0
pyserial==3.5
schedule==1.1.0
